backend/requirements_gradio.txt
gradio==4.10.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
ultralytics==8.0.230
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3
